js: reworked store/reducer patterns
