+++ pkg/interface/src/apps/dojo/app.js
