--- pkg/interface/src/apps/publish/reducers/initial.js
@@ -1,7 +1,0 @@
-import _ from 'lodash';
-
-export class InitialReducer {
-  reduce(json, state) {
-    state.notebooks = json.notebooks || null;
-  }
-}
