--- pkg/interface/src/oldApps/link-webext/browserAction/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <link href="style.css" rel="stylesheet" />
-  </head>
-  <body>
-    <h1 id="myHeading">My browser action</h1>
-    <script src="script.js"></script>
-  </body>
-</html>
