--- pkg/interface/src/oldApps/link-webext/browserAction/script.js
@@ -1,1 +1,0 @@
-console.log('script.js firing');
