--- pkg/interface/src/oldApps/link-webext/options/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <link href="style.css" rel="stylesheet" />
-  </head>
-  <body>
-
-    <form>
-      <label>
-        Ship HTTP endpoint:
-        <input id="endpoint" type="text" placeholder="your-ship.arvo.network" />
-      </label>
-
-      <button type="submit">Save</button>
-    </form>
-
-    <script src="../storage.js"></script>
-    <script src="script.js"></script>
-
-  </body>
-</html>
