--- pkg/interface/src/oldApps/link-webext/options/style.css
@@ -1,3 +1,0 @@
-h1 {
-  font-style: italic;
-}
