--- pkg/interface/src/oldApps/link-webext/storage.js
@@ -1,20 +1,0 @@
-// use synced storage if supported, fall back to local
-const storage = browser.storage.sync || browser.storage.local;
-
-const setEndpoint = (endpoint) => {
-  return storage.set({endpoint});
-}
-
-const getEndpoint = () => {
-  return new Promise((resolve, reject) => {
-    storage.get("endpoint").then((res) => {
-      if (res && res.endpoint) {
-        resolve(res.endpoint);
-      } else {
-        resolve(null);
-      }
-    }, (err) => {
-      resolve(null);
-    });
-  });
-}
