+++ pkg/interface/src/reducers/group-update.js
@@ -1,15 +1,26 @@
 import _ from 'lodash';
 
-export default class GroupUpdateReducer {
+export default class GroupReducer {
+
   reduce(json, state) {
-    const data = _.get(json, 'group-update', false);
+    const data = _.get(json, "group-update", false);
     if (data) {
+      this.initial(data, state);
       this.add(data, state);
       this.remove(data, state);
       this.bundle(data, state);
       this.unbundle(data, state);
       this.keys(data, state);
       this.path(data, state);
+    }
+  }
+
+  initial(json, state) {
+    const data = _.get(json, 'initial', false);
+    if (data) {
+      for (let group in data) {
+       state.groups[group] = new Set(data[group]);
+      }
     }
   }
