+++ pkg/interface/src/reducers/metadata-update.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 
 export default class MetadataReducer {
   reduce(json, state) {
-    const data = _.get(json, 'metadata-update', false);
+    let data = _.get(json, 'metadata-update', false);
     if (data) {
       this.associations(data, state);
       this.add(data, state);
@@ -12,54 +12,60 @@ export default class MetadataReducer {
   }
 
   associations(json, state) {
-    const data = _.get(json, 'associations', false);
+    let data = _.get(json, 'associations', false);
     if (data) {
-      const metadata = state.associations;
-      Object.keys(data).map((channel) => {
-        const channelObj = data[channel];
-        const app = data[channel]['app-name'];
-        if (!(app in metadata)) {
-          metadata[app] = {};
+      let metadata = {};
+      Object.keys(data).forEach((key) => {
+        let val = data[key];
+        let groupPath = val['group-path'];
+        if (!(groupPath in metadata)) {
+          metadata[groupPath] = {};
         }
-        metadata[app][channelObj['app-path']] =  channelObj;
+        metadata[groupPath][key] = val;
       });
       state.associations = metadata;
     }
   }
 
   add(json, state) {
-    const data = _.get(json, 'add', false);
+    let data = _.get(json, 'add', false);
     if (data) {
-      const metadata = state.associations;
-      const app = data['app-name'];
-      if (!(app in metadata)) {
-        metadata[app] = {};
+      let metadata = state.associations;
+      if (!(data['group-path'] in metadata)) {
+        metadata[data['group-path']] = {};
       }
-      metadata[app][data['app-path']] = data;
+      metadata[data['group-path']]
+        [`${data["group-path"]}/${data["app-name"]}${data["app-path"]}`] = data;
+
       state.associations = metadata;
     }
   }
 
   update(json, state) {
-    const data = _.get(json, 'update-metadata', false);
+    let data = _.get(json, 'update-metadata', false);
     if (data) {
-      const metadata = state.associations;
-      const app = data['app-name'];
-      metadata[app][data['app-path']] = data;
+      let metadata = state.associations;
+      if (!(data["group-path"] in metadata)) {
+        metadata[data["group-path"]] = {};
+      }
+      metadata[data["group-path"]][
+        `${data["group-path"]}/${data["app-name"]}${data["app-path"]}`
+      ] = data;
+
       state.associations = metadata;
     }
   }
 
   remove(json, state) {
-    const data = _.get(json, 'remove', false);
+    let data = _.get(json, 'remove', false);
     if (data) {
-      const metadata = state.associations;
-      const app = data['app-name'];
-      if (!(app in metadata)) {
-        return false;
+      let metadata = state.associations;
+      if (data['group-path'] in metadata) {
+        let path =
+        `${data['group-path']}/${data['app-name']}${data['app-path']}`
+        delete metadata[data["group-path"]][path];
+        state.associations = metadata;
       }
-      delete metadata[app][data['app-path']];
-      state.associations = metadata;
     }
   }
 }
